patron_saga/ms-catalogo/requirements.txt
Flask==3.0.3
flask-marshmallow==1.2.1
marshmallow==3.23
Flask-Caching==2.3.0
redis==5.1.1
python-dotenv==1.0.1
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.36
marshmallow-sqlalchemy==1.1.0
Flask-Migrate==4.0.7
psycopg2==2.9.7
psycopg[binary]
Flask-Limiter==3.9.2
uwsgi==2.0.23